A lot of changed on 06.08.23 Variant of login of a lot of activities
