+++ app/src/main/AndroidManifest.xml
@@ -17,6 +17,16 @@
         android:theme="@style/Theme.Moveto"
         tools:targetApi="31">
         <activity
+            android:name="com.maserplay.appai.login.LoginVerifyMailActivity"
+            android:exported="false"
+            android:label="@string/login_login"
+            android:parentActivityName="com.maserplay.appai.login.LoginActivity" />
+        <activity
+            android:name="com.maserplay.appai.login.LoginVerifyActivity"
+            android:exported="false"
+            android:label="@string/login_login"
+            android:parentActivityName="com.maserplay.appai.login.LoginActivity" />
+        <activity
             android:name="com.maserplay.appai.login.LoginActivity"
             android:exported="false"
             android:label="@string/login_login"
