+++ app/src/main/java/com/maserplay/appai/login/LoginAcceptClass.kt
@@ -3,6 +3,14 @@ package com.maserplay.appai.login
 import kotlin.properties.Delegates
 
 class LoginAcceptClass {
-    private var emailaccept by Delegates.notNull<Boolean>()
-    private var totp by Delegates.notNull<Boolean>()
+    var emailaccept: Boolean? = null
+        private set
+    var totp: Boolean? = null
+        private set
+
+    override fun toString(): String {
+        return "LoginAcceptClass(emailaccept=$emailaccept, totp=$totp)"
+    }
+
+
 }
