+++ app/src/main/java/com/maserplay/appai/login/LoginActivity.kt
@@ -1,6 +1,8 @@
 package com.maserplay.appai.login
 
+import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import android.widget.Button
 import android.widget.EditText
@@ -11,34 +13,15 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.widget.addTextChangedListener
 import androidx.lifecycle.ViewModelProvider
 import com.maserplay.AppAi.R
-import org.json.JSONObject
 
 class LoginActivity : AppCompatActivity() {
-    private lateinit var model: LoginViewModel
-    private lateinit var email: EditText
-    private lateinit var password: EditText
-    private lateinit var totp: EditText
-    private lateinit var emailaccept: Button
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_login)
-        email = findViewById(R.id.Email)
-        emailaccept = findViewById(R.id.checkmail)
-        totp = findViewById(R.id.Totp)
-        password = findViewById(R.id.Password)
+        val email: EditText = findViewById(R.id.Email)
+        val password: EditText = findViewById(R.id.Password)
         val loginbutton: Button = findViewById(R.id.login)
-        model = ViewModelProvider(this)[LoginViewModel::class.java]
-
-        findViewById<ScrollView>(R.id.twofact).visibility = model.twofactorll
-        findViewById<ScrollView>(R.id.LoginLayout).visibility = model.loginll
-        findViewById<TextView>(R.id.textView6).visibility = model.totp
-        totp.visibility = model.totp
-        findViewById<TextView>(R.id.textView7).visibility = model.emaillogin
-        emailaccept.visibility = model.emaillogin
-
-        totp.addTextChangedListener {
-            totpaccept()
-        }
+        val model = ViewModelProvider(this)[LoginViewModel::class.java]
         loginbutton.setOnClickListener {
             loginbutton.isEnabled = false
             model.LoginResponseLiveData.observe(this) {
@@ -51,151 +34,36 @@ class LoginActivity : AppCompatActivity() {
 
                     loginbutton.isEnabled = true
                 } else {
-                    val resp: String = it.body().toString()
-                    val json = JSONObject(resp)
+                    val resp = it.body()
                     runOnUiThread {
-                        if (json["email"] == false) {
+                        if (resp?.email == false) {
                             email.error = resources.getString(R.string.login_error_email)
                         }
-                        if (json["password"] == false) {
+                        if (resp?.password == false) {
                             password.error =
                                 resources.getString(R.string.login_error_password)
                             loginbutton.isEnabled = true
                             return@runOnUiThread
                         } else {
-                            if (json["hasaccept"] as Boolean) {
-                                model.cookie = json["acceptcookie"].toString()
-                                model.twofactorll = View.VISIBLE
-                                findViewById<ScrollView>(R.id.twofact).visibility =
-                                    model.twofactorll
-                                model.loginll = View.GONE
-                                findViewById<ScrollView>(R.id.LoginLayout).visibility =
-                                    model.loginll
+                            if (resp?.hasaccept as Boolean) {
+                                model.cookie = resp.acceptcookie.toString()
                                 loginbutton.isEnabled = true
-                                val accept = JSONObject(json["accept"].toString())
-                                if (accept["totp"] as Boolean) {
+                                if (resp.accept?.totp as Boolean) {
                                     model.totp = View.VISIBLE
-                                    findViewById<TextView>(R.id.textView6).visibility =
-                                        model.totp
-                                    totp.visibility =
-                                        model.totp
                                     model.sum--
                                 }
-                                if (accept["emailaccept"] as Boolean) {
+                                if (resp.accept?.emailaccept as Boolean) {
                                     model.emaillogin = View.VISIBLE
-                                    findViewById<TextView>(R.id.textView7).visibility =
-                                        model.emaillogin
-                                    emailaccept.visibility =
-                                        model.emaillogin
                                     model.sum--
                                 }
-                            }
+                                startActivity(Intent(this, LoginVerifyActivity::class.java))
+                            } else { if (resp.verificate == true) { TODO("noverifauth")} else {
+                                startActivity(Intent(this, LoginVerifyMailActivity::class.java))}}
                         }
                     }
                 }
             }
             model.Login(LoginClass(email.text.toString(), password.text.toString()))
         }
-    }
-
-    fun checksum() {
-        if (model.sum >= 0) {
-            findViewById<Button>(R.id.logintwofac).isEnabled = true
-        }
-    }
-
-    fun totpaccept() {
-        if (totp.text.length == 6) {
-            model.LoginAcceptResponseLiveData.observe(this) {
-                if (!it.isSuccessful) {
-                    Toast.makeText(
-                        applicationContext,
-                        getString(R.string.request_error),
-                        Toast.LENGTH_LONG
-                    ).show()
-
-                } else {
-                    val resp: String = it.body().toString()
-                    val json = JSONObject(resp)
-                    if (json["verify"] as Boolean) {
-                        totp.isEnabled = false
-                        model.sum++
-                        checksum()
-                    } else {
-                        totp.error = resources.getString(R.string.login_error_code)
-                    }
-                }
-
-
-            }
-        }
-        model.LoginTotp(LoginVerifySendClass(model.cookie, totp.text.toString().toInt()))
-    }
-
-
-    fun emailaccept(view: View) {
-        model.LoginAcceptResponseLiveData.observe(this) {
-            if (!it.isSuccessful) {
-                Toast.makeText(
-                    applicationContext,
-                    getString(R.string.request_error),
-                    Toast.LENGTH_LONG
-                ).show()
-
-            }
-            val resp: String = it.body().toString()
-            val json = JSONObject(resp)
-            if (json["verify"] as Boolean) {
-                emailaccept.isEnabled = false
-                findViewById<TextView>(R.id.textView8).visibility = View.GONE
-                model.sum++
-                checksum()
-            } else {
-                findViewById<TextView>(R.id.textView8).visibility = View.VISIBLE
-            }
-
-        }
-        model.LoginEmail(LoginVerifySendClass(model.cookie))
-    }
-
-    fun back(view: View) {
-        findViewById<ScrollView>(R.id.twofact).visibility = View.GONE
-        totp.isEnabled = true
-        emailaccept.isEnabled = true
-        model.emaillogin = View.GONE
-        model.totp = View.GONE
-        model.twofactorll = View.GONE
-        model.loginll = View.VISIBLE
-        findViewById<TextView>(R.id.textView6).visibility =
-            View.GONE
-        findViewById<EditText>(R.id.Totp).visibility = View.GONE
-        findViewById<TextView>(R.id.textView7).visibility =
-            View.GONE
-        findViewById<Button>(R.id.checkmail).visibility =
-            View.GONE
-        model.sum = 0
-        totp.setText("")
-        findViewById<TextView>(R.id.textView8).visibility = View.GONE
-        findViewById<ScrollView>(R.id.LoginLayout).visibility = View.VISIBLE
-    }
-
-    fun loginlogin(view: View) {
-        model.LoginResponseLiveData.observe(this) {
-            if (!it.isSuccessful) {
-                Toast.makeText(
-                    applicationContext,
-                    getString(R.string.request_error),
-                    Toast.LENGTH_LONG
-                ).show()
-            } else {
-                val resp: String = it.body().toString()
-                val json = JSONObject(resp)
-                getSharedPreferences("Main", MODE_PRIVATE).edit()
-                    .putString("cookie", json["id"].toString()).apply()
-            }
-        }
-
-        model.LoginCheck(LoginVerifySendClass(model.cookie))
-
     }
 }
