+++ app/src/main/java/com/maserplay/appai/login/LoginVerifyMailActivity.kt
@@ -1,0 +1,12 @@
+package com.maserplay.appai.login
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.maserplay.AppAi.R
+
+class LoginVerifyMailActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_login_verify_mail)
+    }
+}
