+++ app/src/main/res/layout/activity_login_verify_mail.xml
@@ -1,13 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.maserplay.appai.login.LoginActivity"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
+    android:layout_height="match_parent"
+    tools:context="com.maserplay.appai.login.LoginVerifyMailActivity">
 
     <ScrollView
         android:id="@+id/LoginLayout"
@@ -45,39 +42,24 @@
                 android:textSize="20sp"
                 android:textStyle="bold" />
 
-            <EditText
-                android:id="@+id/Email"
+            <TextView
+                android:id="@+id/textView12"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginVertical="10dp"
+                android:layout_height="wrap_content"
                 android:layout_weight="1"
-                android:autofillHints="Email"
-                android:ems="10"
-                android:hint="@string/login_email"
-                android:inputType="textEmailAddress"
-                android:paddingVertical="20dp" />
+                android:gravity="center"
+                android:text="@string/login_verifymail_tittle"
+                android:textSize="34sp" />
 
-            <EditText
-                android:id="@+id/Password"
+            <TextView
+                android:id="@+id/textView13"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginVertical="10dp"
+                android:layout_height="wrap_content"
                 android:layout_weight="1"
-                android:autofillHints="Password"
-                android:ems="10"
-                android:hint="@string/login_password"
-                android:inputType="textPassword"
-                android:paddingVertical="20dp" />
+                android:gravity="center"
+                android:text="@string/login_verifymail_descr"
+                android:textSize="20sp" />
 
-            <Button
-                android:id="@+id/login"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginVertical="10dp"
-                android:layout_weight="1"
-                android:paddingVertical="15dp"
-                android:text="@string/login_login" />
         </LinearLayout>
     </ScrollView>
-
 </androidx.constraintlayout.widget.ConstraintLayout>
