+++ app/src/main/res/values/strings.xml
@@ -66,4 +66,6 @@
     <string name="request_error">An error occurred during the request, check your internet connection and try again later</string>
     <string name="login_error_code">Code not right</string>
     <string name="login_errorinput_mail">Email not verifered</string>
+    <string name="login_verifymail_tittle">Confirm your email</string>
+    <string name="login_verifymail_descr">Registration is completed, but confirmation of your mailing address is required, Check your mailbox.</string>
 </resources>
